styling profile page with progress bar

+++ rak/rak/Models/PointToLevelConverter.h
@@ -3,6 +3,6 @@
 @interface PointToLevelConverter : NSObject
 
 + (NSInteger)getCurrentLevelFromPoints:(NSInteger)points;
-+ (NSInteger)getPercentToNextLevelFromPoints:(NSInteger)points;
++ (float)getPercentToNextLevelFromPoints:(NSInteger)points;
 
 @end
